wrong parametrs of func

+++ tasks/mpi/budazhapova_e_count_freq_character/perf_tests/count_freq_of_ch_perf_tests_mpi.cpp
@@ -16,7 +16,7 @@
 
 namespace budazhapova_e_count_freq_chart_mpi {
 namespace {
-std::string GetRandomString(int length) {
+std::string GetRandomString(long length) {
   static std::string charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890";
   std::string result;
   result.resize(length);
